requirements.txt
python-dotenv==1.0.0
speechrecognition==3.10.0
pyaudio==0.2.13
pyttsx3==2.90
pocketsphinx==5.0.1